+++ temba/orgs/tests/test_org.py
@@ -160,11 +160,11 @@ class OrgTest(TembaTest):
         with self.assertRaises(AssertionError):
             self.org.set_flow_languages(self.admin, ["eng", "eng"])
 
-    def test_country_view(self):
+    def test_locations_view(self):
         self.setUpLocations()
 
         settings_url = reverse("orgs.org_workspace")
-        country_url = reverse("orgs.org_country")
+        country_url = reverse("orgs.org_locations")
 
         rwanda = AdminBoundary.objects.get(name="Rwanda")
 
@@ -176,12 +176,16 @@ class OrgTest(TembaTest):
         response = self.client.get(country_url)
         self.assertEqual(200, response.status_code)
 
+        # clear root_location so we can verify the POST sets it
+        self.org.root_location = None
+        self.org.save(update_fields=("root_location",))
+
         # save with Rwanda as a country
-        self.client.post(country_url, {"country": rwanda.id})
+        self.client.post(country_url, {"root_location": rwanda.id})
 
         # assert it has changed
         self.org.refresh_from_db()
-        self.assertEqual("Rwanda", str(self.org.country))
+        self.assertEqual("Rwanda", str(self.org.root_location))
         self.assertEqual("RW", self.org.default_country_code)
 
         response = self.client.get(settings_url)
@@ -193,18 +197,18 @@ class OrgTest(TembaTest):
             self.assertNotContains(response, "Rwanda")
 
     def test_default_country(self):
-        # if country boundary is set and name is valid country, that has priority
-        self.org.country = AdminBoundary.create(osm_id="171496", name="Ecuador", level=0)
+        # if root location boundary is set and name is valid country, that has priority
+        self.org.root_location = AdminBoundary.create(osm_id="171496", name="Ecuador", level=0)
         self.org.timezone = "Africa/Nairobi"
-        self.org.save(update_fields=("country", "timezone"))
+        self.org.save(update_fields=("root_location", "timezone"))
 
         self.assertEqual("EC", self.org.default_country.alpha_2)
 
         del self.org.default_country
 
         # if country name isn't valid, we'll try timezone
-        self.org.country.name = "Fantasia"
-        self.org.country.save(update_fields=("name",))
+        self.org.root_location.name = "Fantasia"
+        self.org.root_location.save(update_fields=("name",))
 
         self.assertEqual("KE", self.org.default_country.alpha_2)